pyproject.toml
[tool.poetry]
name = "homework-webpython-14"
version = "0.1.0"
description = ""
authors = ["Андрій <andrey.maybrodskyy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.5"
sqlalchemy = "^2.0.7"
alembic = "^1.10.2"
pydantic = {extras = ["email"], version = "^1.10.7"}
fastapi-jwt-auth = "^0.5.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
aioredis = "2.0.0"
fastapi-mail = "^1.2.7"
fastapi-limiter = "^0.1.5"
cloudinary = "^1.32.0"
pytest = "^7.3.1"


[tool.poetry.group.dev.dependencies]
sphinx = "^6.1.3"
pytest-cov = "^4.0.0"
httpx = {extras = ["cli"], version = "^0.24.0"}
pytest-dotenv = "^0.5.2"
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]